pyproject.toml
[tool.poetry]
name = "engagement-agent"
version = "0.1.0"
description = ""
authors = ["shadi-fsai <163780314+shadi-fsai@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
crewai = "^0.55.2"
crewai-tools = "^0.12.0"
discord-py = "^2.4.0"

[tool.poetry.scripts]
Engagement_Agent = "Engagement_Agent.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
